Added a scheduler
I removed the Celery monitor because I read about it more and it doesn't work with MYSQL. It works well with Redis (It scales better and suited for a large dataset and frequent querying).

Added sheduler.py instead. It performs webserver health checks on all of our webservers every 30 seconds and updates the results to the request_history table in our MYSQL database.