Refactor webserver CRUD operations into WebserverService.py & did the same for the scheduler
Moved the CRUD operations logic from the Flask route handler to the WebserverService class.
This change improves the separation of concerns by isolating the business logic from the presentation layer. It enhances the maintainability of the code and makes the service methods reusable across different parts of the application.

Did the same for the scheduler.py - moved the logic to shcedulerService.py